Slide show program

+++ library/graphic/toolkit/vision2-x/pixmap/jpeg/el_jpeg_convertable.e
@@ -6,8 +6,8 @@ note
 	contact: "finnian at eiffel hyphen loop dot com"
 
 	license: "MIT license (See: en.wikipedia.org/wiki/MIT_License)"
-	date: "2022-01-03 15:54:04 GMT (Monday 3rd January 2022)"
-	revision: "7"
+	date: "2022-05-29 14:09:40 GMT (Sunday 29th May 2022)"
+	revision: "8"
 
 deferred class
 	EL_JPEG_CONVERTABLE
@@ -23,12 +23,11 @@ feature -- Basic operations
 	save_as_jpeg (file_path: FILE_PATH; quality: NATURAL)
 		require
 			percentage: 0 <= quality and quality <= 100
-		local
-			surface: like to_pixel_surface
 		do
-			surface := to_pixel_surface
-			surface.save_as_jpeg (file_path, quality)
-			surface.destroy
+			if attached to_pixel_surface as surface then
+				surface.save_as_jpeg (file_path, quality)
+				surface.destroy
+			end
 		end
 
 feature -- Conversion
